Adds stand alone docker image for running ganache-cli, in order to enable connection with metamask
Prepares deno server folder structure

Adds worker to run root cli adapter. Adds read and write through stdin and stdout. Adds tokenizer based on space split of a line and also with minimal support for single and double quotes

Adds decorator for class (cliAdapter) and method (cliEntrypoint) which receive an array of tokens and a description that are associated with calling a specific class and method. Still needs to handle throwing errors. Still needs to understand why is the stding not being read and the tokens interpreted

Fixes issue with stdin not reading, which was actually caused by having the stdin on a worker thread trying to acces an object created on the main thread, ended up not being the same instance of the object. Adds error handling on interpret commands from cli with fallback entrypoints mechanism. Does some overall cleaning and closes the cli lib lib. Still needs to add tests for the cli lib.

Adds some tests for cli lib (not done yet).

All classes decorated with cliAdapter are now singletons. Starts using the singleton to call the class methods

Initial rough (very rough) code for an experimental cli adapter

Implements system of linear equations solving through matrices and gaussian elimination. Compound long using matrices is already created. Still missing the short

Adds short compounding via matrix solving

Creates compoundService and moves compound long matrix code to the service instead of having it directly on the adapter

Moves business code outside of adapter to compoundService. Adds short and long strategies for the inverse of the original pair (i.e., if we only have T1/T2 we can now, know how to compound also T2, buy either using the available inverse pair T2/T1, in case it exists, or use T1/T2 and take profit in terms of T2 instead of just T1)

Adds tokens symbols to the printed messages, to make it easier to understand what to buy and sell

Experimental cli adapter no longer initializes a new compoundService each time a method is called and instead uses the service used in the creation of the adapter instance

Adds custom excel formatting output for the experimental adapater commands

Fixes issue with coefficients calculation for matrixes row on creation of augmented matrices. The cases where the variable was on the right side of a division a wrong coefficient was being considered. Changed to exclude that combination of variables and expressions, as the coefficient in this case of a linear matrix is no possible (or at least no supported by the current system)

Comments out exiting out of solving linear matrices, as it was stopping us from reaching to the point of solving the variables present in the post process equations, which are needed to evaluate the correctness of the solution against all available equations in the end